fix rule registry and correct CNF check method
