+++ antwort/src/main.rs
@@ -1,47 +1,11 @@
-use antwort::macros::register_rule;
-use antwort::rule::RuleApplicationError;
-use antwort::rule_engine::{get_rules, rewrite};
+use antwort::rule_engine::rewrite;
 use antwort::solver::{solve, Clause, Formula};
-use antwort::Expr;
 
-#[register_rule]
-fn example_rule(_expr: &Expr) -> Result<Expr, RuleApplicationError> {
-    Err(RuleApplicationError::RuleNotApplicable)
-}
-
-#[register_rule]
-fn de_morgans1(expr: &Expr) -> Result<Expr, RuleApplicationError> {
-    if let Expr::Negation(b) = expr {
-        if let Expr::Disjunction(exprs) = b.as_ref() {
-            return Ok(Expr::Conjunction(
-                exprs
-                    .iter()
-                    .map(|e| Expr::Negation(Box::new(e.clone())))
-                    .collect(),
-            ));
-        }
-    }
-    Err(RuleApplicationError::RuleNotApplicable)
-}
-
-#[register_rule]
-fn de_morgans2(expr: &Expr) -> Result<Expr, RuleApplicationError> {
-    if let Expr::Negation(b) = expr {
-        if let Expr::Conjunction(exprs) = b.as_ref() {
-            return Ok(Expr::Disjunction(
-                exprs
-                    .iter()
-                    .map(|e| Expr::Negation(Box::new(e.clone())))
-                    .collect(),
-            ));
-        }
-    }
-    Err(RuleApplicationError::RuleNotApplicable)
-}
+// TODO: This should be done via imports, not includes
+include!("./_rules/index.rs");
 
 fn main() {
-    let rules = get_rules();
-    println!("Rules: {:?}", rules);
+    use antwort::Expr;
 
     let expr = Expr::Negation(Box::new(Expr::Disjunction(vec![
         Expr::Variable("a".to_string()),
