--- antwort/tests/integration/demorgans/1.input.cnf.json
@@ -1,14 +1,0 @@
-{
-    "Conjunction": [
-        {
-            "Negation": {
-                "Variable": "a"
-            }
-        },
-        {
-            "Negation": {
-                "Variable": "b"
-            }
-        }
-    ]
-}
